pyproject.toml
[project]
name = "edugraph"
version = "0.1.0"
description = "**EduGraph** is an intelligent learning platform that extracts key topics from uploaded study materials and organizes them into an interactive knowledge graph. It streamlines learning by visually connecting complex concepts, enabling efficient exploration and deeper understanding."
authors = [
    {name = "Ankit Upadhyay"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
boto3 = "^1.36.21"
python-dotenv = "^1.0.1"
neo4j = "^5.28.1"
pypdf = "^5.3.0"
spacy = "^3.8.4"
python-multipart = "^0.0.20"
scikit-learn = "^1.6.1"
pytextrank = "^3.3.0"
llvmlite = "^0.44.0"
bertopic = "^0.16.4"
langchain-core = "^0.3.35"
langchain-text-splitters = "^0.3.6"
langchain-openai = "^0.3.6"
openai = "^1.63.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
